@@ -1,32 +1,39 @@
 {
 	"dependencies": {
-		"@actions/core": "^1.10.1",
-		"finalhandler": "^1.2.0",
-		"node-fetch": "^2.7.0",
-		"puppeteer": "^22",
-		"serve-static": "^1.15.0",
+		"@actions/core": "^1.11.1",
+		"finalhandler": "^2.1.0",
+		"puppeteer": "^24",
+		"serve-static": "^2.2.0",
 		"split2": "^4.2.0",
-		"subresources": "^1.3.3"
+		"subresources": "^2.0.0",
+		"yaml": "^2.8.1"
+	},
+	"pnpm": {
+		"overrides": {
+			"link-checker": "1.4.2",
+			"reffy": "^15",
+			"respec": "^35",
+			"specberus": "^11",
+			"vnu-jar": "^24",
+			"webidl2": "^24"
+		}
 	},
 	"engines": {
-		"node": "^20"
+		"node": "^24"
 	},
-	"packageManager": "pnpm@9.1.3",
+	"type": "module",
+	"packageManager": "pnpm@10.17.0",
 	"scripts": {
-		"build": "tsc -p tsconfig.prod.json",
-		"dev": "tsc -p tsconfig.json -w",
+		"dev": "tsc -w",
+		"typecheck": "tsc",
 		"lint": "prettier . --check"
 	},
 	"devDependencies": {
-		"@types/finalhandler": "^1.2.3",
-		"@types/node": "^20.12.12",
-		"@types/node-fetch": "^2.6.11",
-		"@types/serve-static": "^1.15.7",
+		"@types/finalhandler": "^1.2.4",
+		"@types/node": "^24.5.2",
+		"@types/serve-static": "^1.15.8",
 		"@types/split2": "^4.2.3",
-		"typescript": "^5.4.5"
-	},
-	"optionalDependencies": {
-		"prettier": "^3.2.5",
-		"yaml": "^2.4.2"
+		"prettier": "^3.6.2",
+		"typescript": "^5.9.2"
 	}
 }
