@@ -1,11 +1,11 @@
-import { env, exit, formatAsHeading, pprint, setOutput } from "./utils.js";
+import { env, exit, formatAsHeading, pprint, setOutput } from "./utils.ts";
 
-import { buildOptions } from "./prepare-build.js";
+import { buildOptions } from "./prepare-build.ts";
 import {
 	githubPagesDeployment,
 	w3cEchidnaDeployment,
-} from "./prepare-deploy.js";
-import { validation } from "./prepare-validate.js";
+} from "./prepare-deploy.ts";
+import { validation } from "./prepare-validate.ts";
 
 export interface Inputs {
 	TOOLCHAIN: "respec" | "bikeshed" | string;
@@ -38,7 +38,7 @@ export interface GitHubContext {
 	};
 }
 
-if (module === require.main) {
+if (import.meta.main) {
 	const inputs: Inputs = JSON.parse(env("INPUTS_USER"));
 	const githubContext: GitHubContext = JSON.parse(env("INPUTS_GITHUB"));
 	main(inputs, githubContext).catch(err =>
