@@ -1,11 +1,11 @@
+import * as path from "node:path";
 import { addPath, exportVariable } from "@actions/core";
-import * as path from "path";
-import { ACTION_DIR, PUPPETEER_ENV } from "./constants.js";
-import { env, exit, install, sh } from "./utils.js";
+import { ACTION_DIR, PUPPETEER_ENV } from "./constants.ts";
+import { env, exit, install, sh } from "./utils.ts";
 
 const PYTHONUSERBASE = path.join(ACTION_DIR, "python_modules");
 
-if (module === require.main) {
+if (import.meta.main) {
 	const toolchain = env("INPUTS_TOOLCHAIN");
 	main(toolchain).catch(err => exit(err.message || "Failed", err.code));
 }
