@@ -4,13 +4,12 @@
  *
  * Run this file in a local GitHub repo, and change inputs as needed.
  */
-import { platform } from "os";
-import { formatAsHeading, pprint } from "../src/utils.js";
+import { devNull } from "node:os";
+import { createRequire } from "node:module";
+import { formatAsHeading, pprint } from "../src/utils.ts";
 
 let SILENT_CHILD = !true;
 
-// TODO: use `os.devNull` when switching to Node 16.3+
-const devNull = platform() === "win32" ? String.raw`\\.\nul` : "/dev/null";
 Object.assign(process.env, {
 	GITHUB_ENV: devNull,
 	GITHUB_PATH: devNull,
@@ -20,7 +19,7 @@ const console = global.console;
 if (SILENT_CHILD) {
 	global.console = new Proxy(global.console, {
 		get(target, prop) {
-			// @ts-ignore
+			// @ts-expect-error
 			if (typeof target[prop] === "function") {
 				return () => {};
 			}
@@ -30,9 +29,9 @@ if (SILENT_CHILD) {
 }
 
 export interface Outputs {
-	prepare: Awaited<ReturnType<typeof import("../src/prepare.js").default>>;
-	setup: Awaited<ReturnType<typeof import("../src/setup.js").default>>;
-	build: Awaited<ReturnType<typeof import("../src/build.js").default>>;
+	prepare: Awaited<ReturnType<typeof import("../src/prepare.ts").default>>;
+	setup: Awaited<ReturnType<typeof import("../src/setup.ts").default>>;
+	build: Awaited<ReturnType<typeof import("../src/build.ts").default>>;
 }
 const outputs: Partial<Outputs> = {};
 type AsyncFn = (outputs: Partial<Outputs>) => Promise<object | undefined>;
@@ -46,16 +45,17 @@ const run = (fn: AsyncFn) => async () => {
 	outputs[fn.name as keyof Outputs] = res;
 };
 
+const require = createRequire(import.meta.url);
 Promise.resolve()
-	.then(run(require("./prepare.test.js").default))
-	.then(run(require("./setup.test.js").default))
-	.then(run(require("./validate-input-markup.test.js").default))
-	.then(run(require("./build.test.js").default))
-	.then(run(require("./validate-links.test.js").default))
-	.then(run(require("./validate-markup.test.js").default))
-	.then(run(require("./validate-webidl.test.js").default))
-	.then(run(require("./validate-pubrules.test.js").default))
-	.then(run(require("./deploy-gh-pages.test.js").default))
+	.then(run(require("./prepare.test.ts").default))
+	.then(run(require("./setup.test.ts").default))
+	.then(run(require("./validate-input-markup.test.ts").default))
+	.then(run(require("./build.test.ts").default))
+	.then(run(require("./validate-links.test.ts").default))
+	.then(run(require("./validate-markup.test.ts").default))
+	.then(run(require("./validate-webidl.test.ts").default))
+	.then(run(require("./validate-pubrules.test.ts").default))
+	.then(run(require("./deploy-gh-pages.test.ts").default))
 	.then(() => {
 		console.log();
 		console.log(formatAsHeading("OUTPUTS", "#"));
