@@ -1,5 +1,5 @@
-import main from "../src/setup.js";
-import { Outputs } from "./index.test.js";
+import main from "../src/setup.ts";
+import type { Outputs } from "./index.test.ts";
 
 export default async function setup(outputs: Outputs) {
 	const { toolchain = "respec" } = outputs?.prepare?.build || {};
