@@ -1,5 +1,5 @@
-import main from "../src/validate-input-markup.js";
-import { Outputs } from "./index.test.js";
+import main from "../src/validate-input-markup.ts";
+import type { Outputs } from "./index.test.ts";
 
 export default async function validateInputMarkup(outputs: Outputs) {
 	const { input_markup: shouldValidate = false } =
